Split the build process more clearly between gprbuild and make
Passing the dependency information from gprbuild to Make is
error-prone. We want to call gprbuild, then use Make to detect if the
source or the generator has changed).

Full rebuild seems a bit faster (most of the time is spent executing
the sorter anyway), and now edition an Ada source triggers the rebuild
of its reverse dependencies, including dictionaries.